backend/requirements.txt
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0